A new interface method with an option to choose compress mode when writing bytes.
In the new compression mode, data bytes would be gzipped first and only once, then stored in cassandra by chunks.
When reading bytes, it would try to detected the bytes format with the second chunk's header. As in previous compression mode, each chunk is a complete gzip bytes with a special gzip header.

Signed-off-by: yichen88 <tang.yichenyves@gmail.com>