Remove springfox from dependencies
Springfox does not work with recent spring-boot versions,
and may stop user applications from starting.

Users can restore the configuration on their side if they really
need it.
It was not a good design to include it in the spring module:
applications should be in charge of it.

In any case, the use of springdoc is now encouraged instead.

Signed-off-by: Sylvain Leclerc <sylvain.leclerc@rte-france.com>